faceapi/requirements.dev.txt
SQLAlchemy
typing-extensions
fastapi
pydantic
starlette
email-validator
pydantic-settings
python-multipart
uvicorn[standard]
websockets
sqlmodel
pyjwt
cryptography
passlib
psycopg
psycopg2-binary
requests
numpy
pandas
dlib
face-recognition
face-recognition-models
tqdm
Pillow
scipy
orjson
typish